tests/wpt/tests/css/css-text-decor/parsing/text-decoration-style-valid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Text Decoration Test: Parsing text-decoration-style with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-3/#text-decoration-style-property">
<meta name="assert" content="text-decoration-style supports the full grammar 'solid | double | dotted | dashed | wavy'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_valid_value("text-decoration-style", "solid");
test_valid_value("text-decoration-style", "double");
test_valid_value("text-decoration-style", "dotted");
test_valid_value("text-decoration-style", "dashed");
test_valid_value("text-decoration-style", "wavy");
</script>
